backend/requirements.txt
Flask==3.1.1
flask-cors==6.0.1
supabase==2.16.0
python-jose[cryptography]==3.5.0
bcrypt==4.3.0
requests==2.32.4
pillow==11.3.0
python-dotenv==1.1.1

